Continue service refactoring Moved parser-router to a service
