+++ services/parser-router/sls-functions.yml
@@ -1,0 +1,17 @@
+s3-event-handler:
+  handler: src/fileRouterLambda.handle
+  environment:
+    ulster_csv: ${self:custom.ulsterCsvParserSQSQueueUrl}
+  events:
+    - s3:
+        bucket: ${self:custom.userFilesBucket}
+        event: s3:ObjectCreated:*
+        rules:
+          - prefix: account-files/
+        existing: true
+  iamRoleStatementsName: virtwallet-parser-router-${self:provider.stage}-iam-role
+  iamRoleStatements:
+    - Effect: Allow
+      Action: 
+        - sqs:SendMessage
+      Resource: ${self:custom.ulsterCsvParserSQSQueueArn}
